scripts: Makefile: disable size_append
techpack/audio is somehow triggering size_append
which is time-consuming, resource-hungry
and utterly meaningless during incremental builds.

Likely there are BUGs in Kbuild but so far I have
not been able to find them.

As we are not using size_append for this kernel,
let's just disable size_append before a proper
solution can be found.

Change-Id: I9ba68da339629a71cf7a8a7c0f7046fc11f14582
Signed-off-by: Jesse Chan <jc@linux.com>